requirements.txt
# Core voice pipeline
pipecat-ai

# Networking and server
fastapi
uvicorn
websockets
aiohttp

# STT/TTS integration
nemo_toolkit[asr]  # Parakeet streaming (optional, GPU recommended)
soundfile
numpy

# LLM integration
ollama  # Python client for the Ollama API
pydantic
python-dotenv

# Tooling
tavily-python  # Web search API
duckdb  # Lightweight DB for memory store
pyyaml

# Development utilities
pytest
